translate/requirements.txt
# requirements.txt
# 用于日语视频转中文字幕的自动化工具

ffmpeg-python==0.2.0
whisper==1.1.10
pysrt
tqdm
openai